requirements.txt
discord.py==2.3.2
aiohttp==3.9.4
python-dotenv==1.0.1
